Update for a build fix.

+++ src/srtla-source.c
@@ -173,7 +173,7 @@ static void srtla_audio_capture_cb(void *param, obs_source_t *source, const stru
 	for (uint32_t c = 0; c < out.speakers; c++) {
 		if (out.data[c]) {
 			float *ch_data = (float *)out.data[c];
-			for (uint32_t i = 0; i < out.frames; i++) {
+			for (uint32_t i = 0; i < (uint32_t)out.frames; i++) {
 				float val = fabsf(ch_data[i]);
 				if (val > max_peak) max_peak = val;
 			}
